Updated logging to remove unsigned bug
Changed all rms_id fields to "long" from "unsigned long" to eliminate
a bug where signed values were being printed as unsigned values. Also
changed appropriate logging code.

+++ include/linux/sched.h
@@ -52,7 +52,7 @@ struct sched_param {
 	int sched_priority;
 
 #ifdef	CONFIG_SCHED_RMS_POLICY
-	unsigned int	rms_id;
+	long	rms_id;
 	unsigned long long period;
 #endif
 
@@ -1235,7 +1235,7 @@ struct rcu_node;
 struct task_struct {
 	
 #ifdef	CONFIG_SCHED_RMS_POLICY
-	unsigned int	rms_id;
+	long	rms_id;
 	unsigned long long period;
 #endif
